+++ CMakeLists.txt
@@ -7,8 +7,20 @@ set(THREADS_PREFER_PTHREAD_FLAG ON)
 
 find_package(Threads REQUIRED)
 
-SET(SOURCES "adaptive/adaptive.cpp")
-SET(CONNECTOR_SOURCES "adaptive/adaptive_c_connector.cpp")
+SET(SOURCES 
+    "adaptive/adaptive.hpp"
+    "adaptive/_defines.hpp"
+    "adaptive/_parallel_for_worker.hpp"
+    "adaptive/_reduction_worker.hpp"
+    "adaptive/_thread_handler.cpp"
+    "adaptive/_thread_handler.hpp"
+    "adaptive/_worker.hpp"
+    "adaptive/atomic_barrier.cpp"
+    "adaptive/atomic_barrier.hpp"
+    "adaptive/atomic_mutex.cpp"
+    "adaptive/atomic_mutex.hpp"
+    )
+SET(CONNECTOR_SOURCES "adaptive/adaptive_c_connector.cpp" "adaptive/adaptive.h")
 
 ADD_LIBRARY(adaptive SHARED ${SOURCES} ${CONNECTOR_SOURCES})
 target_link_libraries(adaptive Threads::Threads)
@@ -24,6 +36,15 @@ install(
         COMPONENT Libraries
 )
 install(
-    FILES "adaptive/adaptive.hpp" "adaptive/adaptive.h"
-        DESTINATION "include/adaptive"
+    FILES 
+        "adaptive/adaptive.hpp"
+        "adaptive/adaptive.h"
+        "adaptive/_defines.hpp"
+        "adaptive/_parallel_for_worker.hpp"
+        "adaptive/_reduction_worker.hpp"
+        "adaptive/_thread_handler.hpp"
+        "adaptive/_worker.hpp"
+        "adaptive/atomic_barrier.hpp"
+        "adaptive/atomic_mutex.hpp"
+    DESTINATION "include/adaptive"
 )
